requirements.txt
agate==1.6.3
agate-dbf==0.2.2
agate-excel==0.2.5
agate-sql==0.5.8
aiogram==2.22.1
aiohttp==3.8.1
aioschedule==0.5.2
aiosignal==1.2.0
asttokens==2.0.8
async-generator==1.10
async-timeout==4.0.2
attrs==22.1.0
Babel==2.9.1
backcall==0.2.0
beautifulsoup4==4.11.1
biopython==1.79
certifi==2022.6.15
charset-normalizer==2.1.1
csvkit==1.0.7
dbfread==2.0.7
debugpy==1.6.3
decorator==5.1.1
editdistpy==0.1.3
emoji==2.0.0
emoji-country-flag==1.3.1
enforce-typing==1.0.0.post1
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.1.1
Flag==0.1.1
free-proxy==1.0.6
frozenlist==1.3.1
future==0.18.2
google-search-results==2.4.1
greenlet==1.1.3
h11==0.13.0
idna==3.3
ipykernel==6.16.0
ipython==8.5.0
isodate==0.6.1
jedi==0.18.1
jupyter-core==4.11.1
jupyter_client==7.4.2
leather==0.3.4
lxml==4.9.1
matplotlib-inline==0.1.6
multidict==6.0.2
nest-asyncio==1.5.6
numpy==1.23.3
olefile==0.46
openpyxl==3.0.10
outcome==1.2.0
packaging==21.3
parsedatetime==2.4
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.31
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.13.0
pyparsing==3.0.9
PySocks==1.7.1
pyspellchecker==0.7.0
python-dateutil==2.8.2
python-dotenv==0.21.0
python-slugify==6.1.2
pytimeparse==1.1.8
pytrials==0.2.0
pytz==2022.2.1
pyzmq==24.0.1
requests==2.28.1
selenium==4.4.3
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.41
stack-data==0.5.1
text-unidecode==1.3
tornado==6.2
tqdm==4.64.1
traitlets==5.4.0
trio==0.21.0
trio-websocket==0.9.2
urllib3==1.26.12
wcwidth==0.2.5
webdriver-manager==3.8.3
wsproto==1.2.0
xlrd==2.0.1
yarl==1.8.1
